django_admin/requirements.txt
asgiref==3.5.0
Django==3.2.6
Faker==13.3.4
flake8==4.0.0
mccabe==0.6.1
python-dotenv==0.20.0
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
split_settings
gunicorn==20.1.0
sentry-sdk==1.9.0